update_premium_pattern_wrapper works with v3 data
Function that fetches premium or enterprise data files has been updated.
Important changes are:
- Format of the request string has been changed. A binary data file is
now requested instead of the Python specific device data. This is due to
changes in version 3 that allow all platforms to use the same data file.
- The update process itself has been changed. Once file is downloaded it
is no longer installed via pip install. Instead it is uncompressed and
copied in place of the file speciffied in the V3_WRAPPER_DATABASE.

Important: auto-update does not currently work with Trie data files.